Move old API under specific package version
Prepare space for new versions so that old code wouldn't need to be
modified.
