+++ src/main/java/uk/gov/hmcts/reform/pdf/service/endpoint/v1/PDFGenerationEndpoint.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pdf.service.endpoint;
+package uk.gov.hmcts.reform.pdf.service.endpoint.v1;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
